Support multi-character constants with GCC-style byte packing
A character constant with more than one character (e.g. 'ab') was
silently decoded to its first character only. Pack its bytes instead,
GCC-style: big-endian, zero-padded from the left.

- A byte with bit7=0 is a single ASCII byte; a byte with bit7=1 must
  begin a valid UTF-8 sequence (lead + continuation bytes validated),
  whose raw bytes are kept verbatim. A backslash escape contributes its
  byte value (low 8 bits) with no UTF-8 check. Invalid UTF-8 or more than
  six packed bytes is a fatal error.
- Result type: 1-5 bytes -> int; exactly 6 bytes (48 bits) -> unsigned
  int. The 6-byte unsignedness deviates from the standard and is a
  deliberate extension, documented in README.

To carry values up to 48 bits through a host pipeline that backed int
constants with 32-bit storage, widen Literal/Tac_Const int_val/uint_val
to int64_t/uint64_t (also forward-looking for backends whose int exceeds
32 bits): update val_int/val_uint signatures, non-truncating
serialization casts, and 64-bit print/yaml/graphviz format specifiers.
An int static initializer now uses the 48-bit-capable INIT_I64 slot
(mirrors the existing TYPE_UINT -> INIT_U64 choice; emits identically on
BESM-6 - one word, masked to 41/48 bits).

Tests: parser value/kind and death tests, scanner multi-char lexeme, TAC
binary round-trip of wide int/uint constants, and a BESM-6 end-to-end run
(static global + inline constant). Golden tests updated for INIT_I64.

+++ ast/ast.h
@@ -7,6 +7,7 @@ extern "C" {
 
 #include <stdbool.h>
 #include <stddef.h>
+#include <stdint.h>
 #include <stdio.h>
 
 /* Forward declarations for recursive types */
@@ -390,10 +391,10 @@ typedef enum {
 struct Literal {
     LiteralKind kind;
     union {
-        int int_val;
+        int64_t int_val;
         long long_val;
         long long long_long_val;
-        unsigned int uint_val;
+        uint64_t uint_val;
         unsigned long ulong_val;
         unsigned long long ulong_long_val;
         double real_val;
